research/embedding_vs_finetuning/requirements.txt
# Research: Embedding vs Fine-Tuning Experiment
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.1.0
transformers>=4.35.0
numpy>=1.26.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Embeddings
sentence-transformers>=2.2.2

# Gradient boosting
lightgbm>=4.1.0

# Report generation
python-docx>=1.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Data handling
pandas>=2.1.0
pyyaml>=6.0.1
tqdm>=4.66.0
